+++ config/config.go
@@ -10,8 +10,8 @@ import (
 	"time"
 
 	"github.com/Shopify/sarama"
+	"github.com/mailgun/kazoo-go"
 	"github.com/pkg/errors"
-	"github.com/wvanbergen/kazoo-go"
 	"gopkg.in/yaml.v2"
 )
 
@@ -376,8 +376,6 @@ func (p *Proxy) validate() error {
 		return errors.New("consumer.offsets_commit_interval must be > 0")
 	case p.Consumer.OffsetsCommitTimeout <= 0:
 		return errors.New("consumer.offsets_commit_timeout must be > 0")
-	case p.Consumer.RebalanceDelay <= 0:
-		return errors.New("consumer.rebalance_delay must be > 0")
 	case p.Consumer.SubscriptionTimeout <= 0:
 		return errors.New("consumer.subscription_timeout must be > 0")
 	case p.Consumer.RetryBackoff <= 0:
@@ -419,7 +417,7 @@ func defaultProxyWithClientID(clientID string) *Proxy {
 	c.Producer.RetryMax = 6
 	c.Producer.ShutdownTimeout = 30 * time.Second
 
-	c.Consumer.AckTimeout = 15 * time.Second
+	c.Consumer.AckTimeout = 300 * time.Second
 	c.Consumer.ChannelBufferSize = 64
 	c.Consumer.FetchMaxBytes = 1024 * 1024
 	c.Consumer.FetchMaxWait = 250 * time.Millisecond
@@ -428,8 +426,7 @@ func defaultProxyWithClientID(clientID string) *Proxy {
 	c.Consumer.MaxRetries = -1
 	c.Consumer.OffsetsCommitInterval = 500 * time.Millisecond
 	c.Consumer.OffsetsCommitTimeout = 1500 * time.Millisecond
-	c.Consumer.RebalanceDelay = 250 * time.Millisecond
-	c.Consumer.SubscriptionTimeout = 20 * time.Second
+	c.Consumer.SubscriptionTimeout = 15 * time.Second
 	c.Consumer.RetryBackoff = 500 * time.Millisecond
 	return c
 }