+++ consumer/subscriber/subscriber_test.go
@@ -1,6 +1,7 @@
 package subscriber
 
 import (
+	"reflect"
 	"sync"
 	"testing"
 	"time"
@@ -9,7 +10,7 @@ import (
 	"github.com/mailgun/kafka-pixy/config"
 	"github.com/mailgun/kafka-pixy/none"
 	"github.com/mailgun/kafka-pixy/testhelpers"
-	"github.com/wvanbergen/kazoo-go"
+	"github.com/mailgun/kazoo-go"
 	. "gopkg.in/check.v1"
 )
 
@@ -17,63 +18,27 @@ func Test(t *testing.T) {
 	TestingT(t)
 }
 
-type GroupMemberSuite struct {
+type SubscriberSuite struct {
 	ns       *actor.Descriptor
 	kazooClt *kazoo.Kazoo
 }
 
-var _ = Suite(&GroupMemberSuite{})
+var _ = Suite(&SubscriberSuite{})
 
-func (s *GroupMemberSuite) SetUpSuite(c *C) {
+func (s *SubscriberSuite) SetUpSuite(c *C) {
 	testhelpers.InitLogging()
 	var err error
 	s.kazooClt, err = kazoo.NewKazoo(testhelpers.ZookeeperPeers, kazoo.NewConfig())
 	c.Assert(err, IsNil)
 }
 
-func (s *GroupMemberSuite) SetUpTest(c *C) {
+func (s *SubscriberSuite) SetUpTest(c *C) {
 	s.ns = actor.Root().NewChild("T")
-}
-
-func (s *GroupMemberSuite) TestNormalizeTopics(c *C) {
-	c.Assert(normalizeTopics(nil), DeepEquals, []string(nil))
-	c.Assert(normalizeTopics([]string{}), DeepEquals, []string(nil))
-	c.Assert(normalizeTopics([]string{"c", "a", "b"}), DeepEquals, []string{"a", "b", "c"})
-
-	c.Assert(normalizeTopics([]string{"c", "a", "b"}), Not(DeepEquals), []string{"a", "b"})
-}
-
-func (s *GroupMemberSuite) TestTopicsEqual(c *C) {
-	c.Assert(topicsEqual([]string{}, nil), Equals, true)
-	c.Assert(topicsEqual(nil, []string{}), Equals, true)
-	c.Assert(topicsEqual([]string{}, []string{}), Equals, true)
-	c.Assert(topicsEqual([]string{"a"}, []string{"a"}), Equals, true)
-	c.Assert(topicsEqual([]string{"a", "b", "c"}, []string{"a", "b", "c"}), Equals, true)
-
-	c.Assert(topicsEqual([]string{"a", "b", "c"}, []string{"a", "b"}), Equals, false)
-	c.Assert(topicsEqual([]string{"a", "b"}, []string{"b", "a"}), Equals, false)
-}
-
-func (s *GroupMemberSuite) TestSubscriptionsEqual(c *C) {
-	c.Assert(subscriptionsEqual(nil, nil), Equals, true)
-	c.Assert(subscriptionsEqual(map[string][]string{}, nil), Equals, true)
-	c.Assert(subscriptionsEqual(nil, map[string][]string{}), Equals, true)
-	c.Assert(subscriptionsEqual(map[string][]string{}, map[string][]string{}), Equals, true)
-
-	c.Assert(subscriptionsEqual(
-		map[string][]string{
-			"m1": {"a", "b"},
-			"m2": {"c", "d", "e"},
-		},
-		map[string][]string{
-			"m1": {"a", "b"},
-			"m2": {"c", "d", "e"},
-		}), Equals, true)
 }
 
 // When a list of topics is sent to the `topics()` channel, a membership change
 // is received with the same list of topics for the registrator name.
-func (s *GroupMemberSuite) TestSimpleSubscribe(c *C) {
+func (s *SubscriberSuite) TestSimpleSubscribe(c *C) {
 	cfg := newConfig("m1")
 	ss := Spawn(s.ns.NewChild("m1"), "g1", cfg, s.kazooClt)
 	defer ss.Stop()
@@ -89,7 +54,7 @@ func (s *GroupMemberSuite) TestSimpleSubscribe(c *C) {
 // When topic subscription changes occur in close succession only one
 // membership change notification is received back with the most recent topic
 // list for the registrator name.
-func (s *GroupMemberSuite) TestSubscribeSequence(c *C) {
+func (s *SubscriberSuite) TestSubscribeSequence(c *C) {
 	cfg := newConfig("m1")
 	ss := Spawn(s.ns.NewChild("m1"), "g1", cfg, s.kazooClt)
 	defer ss.Stop()
@@ -103,9 +68,9 @@ func (s *GroupMemberSuite) TestSubscribeSequence(c *C) {
 		map[string][]string{"m1": {"bazz", "blah"}})
 }
 
-// If a group member resubscribes to the same list of topics, then nothing is
-// updated.
-func (s *GroupMemberSuite) TestReSubscribe(c *C) {
+// If a group member resubscribes to the same list of topics, then the same
+// member subscriptions are returned.
+func (s *SubscriberSuite) TestReSubscribe(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -120,24 +85,19 @@ func (s *GroupMemberSuite) TestReSubscribe(c *C) {
 		"m1": {"bar", "foo"},
 		"m2": {"bar", "bazz"},
 	}
-	c.Assert(<-ss1.Subscriptions(), DeepEquals, membership)
-	c.Assert(<-ss2.Subscriptions(), DeepEquals, membership)
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 
 	// When
 	ss1.Topics() <- []string{"foo", "bar"}
 
 	// Then
-	select {
-	case update := <-ss1.Subscriptions():
-		c.Errorf("Unexpected update: %v", update)
-	case update := <-ss2.Subscriptions():
-		c.Errorf("Unexpected update: %v", update)
-	case <-time.After(300 * time.Millisecond):
-	}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 }
 
 // To unsubscribe from all topics an empty topic list can be sent.
-func (s *GroupMemberSuite) TestSubscribeToNothing(c *C) {
+func (s *SubscriberSuite) TestSubscribeToNothing(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -146,23 +106,21 @@ func (s *GroupMemberSuite) TestSubscribeToNothing(c *C) {
 	defer ss2.Stop()
 	ss1.Topics() <- []string{"foo", "bar"}
 	ss2.Topics() <- []string{"foo"}
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}})
-	c.Assert(<-ss2.Subscriptions(), DeepEquals,
-		map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}})
+	membership := map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 
 	// When
 	ss1.Topics() <- []string{}
 
 	// Then
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{"m2": {"foo"}})
-	c.Assert(<-ss2.Subscriptions(), DeepEquals,
-		map[string][]string{"m2": {"foo"}})
+	membership = map[string][]string{"m2": {"foo"}}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 }
 
 // To unsubscribe from all topics nil value can be sent.
-func (s *GroupMemberSuite) TestSubscribeToNil(c *C) {
+func (s *SubscriberSuite) TestSubscribeToNil(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -171,34 +129,30 @@ func (s *GroupMemberSuite) TestSubscribeToNil(c *C) {
 	defer ss2.Stop()
 	ss1.Topics() <- []string{"foo", "bar"}
 	ss2.Topics() <- []string{"foo"}
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}})
-	c.Assert(<-ss2.Subscriptions(), DeepEquals,
-		map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}})
+	membership := map[string][]string{"m1": {"bar", "foo"}, "m2": {"foo"}}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 
 	// When
 	ss1.Topics() <- nil
 
 	// Then
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{"m2": {"foo"}})
-	c.Assert(<-ss2.Subscriptions(), DeepEquals,
-		map[string][]string{"m2": {"foo"}})
+	membership = map[string][]string{"m2": {"foo"}}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
 }
 
 // It is possible to subscribe to a non-empty list of topics after
 // unsubscribing from everything.
-func (s *GroupMemberSuite) TestSomethingAfterNothingBug(c *C) {
+func (s *SubscriberSuite) TestSomethingAfterNothingBug(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
 
 	ss1.Topics() <- []string{"foo"}
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{"m1": {"foo"}})
+	c.Assert(<-ss1.Subscriptions(), DeepEquals, map[string][]string{"m1": {"foo"}})
 	ss1.Topics() <- []string{}
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{})
+	c.Assert(<-ss1.Subscriptions(), DeepEquals, map[string][]string{})
 
 	// When
 	ss1.Topics() <- []string{"foo"}
@@ -211,7 +165,7 @@ func (s *GroupMemberSuite) TestSomethingAfterNothingBug(c *C) {
 // When several different registrator instances subscribe to the same group,
 // they all receive identical membership change notifications that include all
 // their subscription.
-func (s *GroupMemberSuite) TestMembershipChanges(c *C) {
+func (s *SubscriberSuite) TestMembershipChanges(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -233,15 +187,17 @@ func (s *GroupMemberSuite) TestMembershipChanges(c *C) {
 		"m2": {"foo"},
 		"m3": {"bazz", "blah", "foo"}}
 
-	c.Assert(<-ss1.Subscriptions(), DeepEquals, membership)
-	c.Assert(<-ss2.Subscriptions(), DeepEquals, membership)
-	c.Assert(<-ss3.Subscriptions(), DeepEquals, membership)
+	assertSubscription(c, ss1.Subscriptions(), membership, 5*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 5*time.Second)
+	assertSubscription(c, ss3.Subscriptions(), membership, 5*time.Second)
 }
 
-// When one of the group members generates a rapid sequence of subscription
-// changes so that at the end its subscription is the same as in the beginning
-// of the sequence then other members won't be notified of such changes.
-func (s *GroupMemberSuite) TestRedundantUpdateIgnored(c *C) {
+// Redundant updates used to be ignored, but that turned out to be wrong. Due
+// to the ZooKeeper single-fire watch semantic it is possible to miss
+// intermediate changes and only see the final subscription state, which make
+// look the same as the last a subscriber had seen and be ignored. But topic
+// consumers could have been changed and in need of rewiring.
+func (s *SubscriberSuite) TestRedundantUpdateBug(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -252,25 +208,21 @@ func (s *GroupMemberSuite) TestRedundantUpdateIgnored(c *C) {
 	ss1.Topics() <- []string{"foo", "bar"}
 	ss2.Topics() <- []string{"foo", "bazz", "blah"}
 
-	c.Assert(<-ss1.Subscriptions(), DeepEquals,
-		map[string][]string{
-			"m1": {"bar", "foo"},
-			"m2": {"bazz", "blah", "foo"}})
+	membership := map[string][]string{
+		"m1": {"bar", "foo"},
+		"m2": {"bazz", "blah", "foo"}}
+	c.Assert(<-ss1.Subscriptions(), DeepEquals, membership)
 
 	// When
 	ss2.Topics() <- []string{"bar"}
 	ss2.Topics() <- []string{"foo", "bazz", "blah"}
 
 	// Then
-	select {
-	case update := <-ss1.Subscriptions():
-		c.Errorf("Unexpected update: %v", update)
-	case <-time.After(300 * time.Millisecond):
-	}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
 }
 
 // When a group registrator claims a topic partitions it becomes its owner.
-func (s *GroupMemberSuite) TestClaimPartition(c *C) {
+func (s *SubscriberSuite) TestClaimPartition(c *C) {
 	cfg := newConfig("m1")
 	ss := Spawn(s.ns.NewChild("m1"), "g1", cfg, s.kazooClt)
 	defer ss.Stop()
@@ -292,7 +244,7 @@ func (s *GroupMemberSuite) TestClaimPartition(c *C) {
 
 // If a consumer group member instance tries to acquire a partition that has
 // already been acquired by another member then it fails.
-func (s *GroupMemberSuite) TestClaimPartitionClaimed(c *C) {
+func (s *SubscriberSuite) TestClaimPartitionClaimed(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -318,7 +270,7 @@ func (s *GroupMemberSuite) TestClaimPartitionClaimed(c *C) {
 }
 
 // It is ok to claim the same partition twice by the same group member.
-func (s *GroupMemberSuite) TestClaimPartitionTwice(c *C) {
+func (s *SubscriberSuite) TestClaimPartitionTwice(c *C) {
 	cfg := newConfig("m1")
 	ss := Spawn(s.ns.NewChild("m1"), "g1", cfg, s.kazooClt)
 	defer ss.Stop()
@@ -338,7 +290,7 @@ func (s *GroupMemberSuite) TestClaimPartitionTwice(c *C) {
 
 // If a partition has been claimed more then once then it is release as soon as
 // any of the claims is revoked.
-func (s *GroupMemberSuite) TestReleasePartition(c *C) {
+func (s *SubscriberSuite) TestReleasePartition(c *C) {
 	cfg := newConfig("m1")
 	ss := Spawn(s.ns.NewChild("m1"), "g1", cfg, s.kazooClt)
 	defer ss.Stop()
@@ -359,7 +311,7 @@ func (s *GroupMemberSuite) TestReleasePartition(c *C) {
 
 // If a partition is claimed by another group member then `ClaimPartition` call
 // blocks until it is released.
-func (s *GroupMemberSuite) TestClaimPartitionParallel(c *C) {
+func (s *SubscriberSuite) TestClaimPartitionParallel(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -386,7 +338,7 @@ func (s *GroupMemberSuite) TestClaimPartitionParallel(c *C) {
 
 // If a partition is claimed by another group member then `ClaimPartition` call
 // blocks until it is released.
-func (s *GroupMemberSuite) TestClaimPartitionCanceled(c *C) {
+func (s *SubscriberSuite) TestClaimPartitionCanceled(c *C) {
 	cfg1 := newConfig("m1")
 	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
 	defer ss1.Stop()
@@ -426,6 +378,56 @@ func (s *GroupMemberSuite) TestClaimPartitionCanceled(c *C) {
 	wg.Wait()
 }
 
+// If claiming a partition fails then the subscription is updated to make all
+// group members re-read the entire group subscription state.
+func (s *SubscriberSuite) TestClaimClaimed(c *C) {
+	cfg1 := newConfig("m1")
+	cfg1.Consumer.RetryBackoff = 150 * time.Millisecond
+	ss1 := Spawn(s.ns.NewChild("m1"), "g1", cfg1, s.kazooClt)
+	defer ss1.Stop()
+	cfg2 := newConfig("m2")
+	cfg2.Consumer.RetryBackoff = 150 * time.Millisecond
+	ss2 := Spawn(s.ns.NewChild("m2"), "g1", cfg2, s.kazooClt)
+	defer ss2.Stop()
+	cfg3 := newConfig("m3")
+	cfg3.Consumer.RetryBackoff = 150 * time.Millisecond
+	ss3 := Spawn(s.ns.NewChild("m3"), "g1", cfg3, s.kazooClt)
+	defer ss3.Stop()
+	cancelCh := make(chan none.T)
+	wg := &sync.WaitGroup{}
+
+	ss1.Topics() <- []string{"foo", "bar"}
+	ss2.Topics() <- []string{"foo", "bazz", "blah"}
+	ss3.Topics() <- []string{"bazz"}
+
+	membership := map[string][]string{
+		"m1": {"bar", "foo"},
+		"m2": {"bazz", "blah", "foo"},
+		"m3": {"bazz"}}
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss3.Subscriptions(), membership, 3*time.Second)
+
+	claim1 := ss1.ClaimPartition(s.ns, "foo", 1, cancelCh)
+	defer claim1()
+
+	// When: try to claim a partition that is claimed by another member.
+	go func() {
+		ss2.ClaimPartition(s.ns, "foo", 1, cancelCh)()
+	}()
+
+	// Then until the retry backoff timeout is not elapsed we got nothing.
+	assertNoneReceived(c, ss3.Subscriptions(), 100*time.Millisecond)
+	// After the retry backoff timeout is elapsed all members get their
+	assertSubscription(c, ss1.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss2.Subscriptions(), membership, 3*time.Second)
+	assertSubscription(c, ss3.Subscriptions(), membership, 3*time.Second)
+
+	close(cancelCh)
+	// Wait for all test goroutines to stop.
+	wg.Wait()
+}
+
 // partitionOwner returns the id of the consumer group member that has claimed
 // the specified topic/partition.
 func partitionOwner(gm *T, topic string, partition int32) (string, error) {
@@ -442,6 +444,27 @@ func partitionOwner(gm *T, topic string, partition int32) (string, error) {
 func newConfig(clientId string) *config.Proxy {
 	cfg := config.DefaultProxy()
 	cfg.ClientID = clientId
-	cfg.Consumer.RebalanceDelay = 100 * time.Millisecond
 	return cfg
+}
+
+func assertSubscription(c *C, ch <-chan map[string][]string, want map[string][]string, timeout time.Duration) {
+	for {
+		select {
+		case got := <-ch:
+			if reflect.DeepEqual(got, want) {
+				return
+			}
+		case <-time.After(timeout):
+			c.Errorf("Timeout waiting for %v", want)
+			return
+		}
+	}
+}
+
+func assertNoneReceived(c *C, ch <-chan map[string][]string, timeout time.Duration) {
+	select {
+	case got := <-ch:
+		c.Errorf("Unexpected update %v", got)
+	case <-time.After(timeout):
+	}
 }